+++ src/Services/ServiceType.php
@@ -4,6 +4,7 @@ namespace DreamFactory\Core\Services;
 
 use DreamFactory\Core\Contracts\ServiceConfigHandlerInterface;
 use DreamFactory\Core\Contracts\ServiceTypeInterface;
+use DreamFactory\Core\Enums\VerbsMask;
 
 /**
  * Interface ServiceType
@@ -55,6 +56,10 @@ class ServiceType implements ServiceTypeInterface
      * @var callable Designated callback for creating a service of this type
      */
     protected $factory = null;
+    /**
+     * @var callable Designated role access exceptions for handling service of this type
+     */
+    protected $accessExceptions = null;
 
     /**
      * Create a new service type instance.
@@ -125,6 +130,25 @@ class ServiceType implements ServiceTypeInterface
     public function make($name, array $config = [])
     {
         return call_user_func($this->factory, $config, $name);
+    }
+
+    public function isAccessException($action, $path = null)
+    {
+        if (is_string($action)) {
+            $action = VerbsMask::toNumeric($action);
+        }
+        if (isset($this->accessExceptions)) {
+            foreach ($this->accessExceptions as $exception) {
+                if (($action & array_get($exception, 'verb_mask')) &&
+                    (('*' === array_get($exception, 'resource')) ||
+                        ($path === array_get($exception, 'resource')))
+                ) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
     }
 
     public function toArray()